Support GSM AT commands for SMS over bluetooth (1/2)
Modifications required to support the older GSM "AT+" commands for SMS over
the bluetooth transport for use with car radios, etc.  This must piggy back
on the existing handsfree code as the spec uses that channel for the commands.
(2 parts: frameworks/base and packages/apps/Phone)

frameworks/base:
Additions to JNI and supporting framework code to allow sending of individual
characters and to receive the PDU stream - neither of which are CR terminated.

Change-Id: Ib0d0a2ba1ffcd55708b42f53ff38c6ff4a7292d1
Signed-off-by: Gary Dezern <garyd9@gmail.com>